+++ homework4_1/pom.xml
@@ -49,20 +49,15 @@
             <version>1.2</version>
         </dependency>
 
-
         <!-- https://mvnrepository.com/artifact/com.h2database/h2 -->
         <dependency>
             <groupId>com.h2database</groupId>
             <artifactId>h2</artifactId>
             <version>1.4.200</version>
         </dependency>
-
-
-
-
-
-
     </dependencies>
+
+
 
 
     <build>
