+++ homework4_1/src/main/java/com/sergio/Config.java
@@ -1,0 +1,4 @@
+package com.sergio;
+
+public class Config {
+}
