+++ homework4_1/src/main/java/com/sergio/Listener.java
@@ -2,7 +2,6 @@
 package com.sergio;
 
 import com.sergio.sql.SqlHelper;
-
 import javax.servlet.ServletContextEvent;
 import javax.servlet.ServletContextListener;
 import javax.servlet.annotation.WebListener;
@@ -14,12 +13,8 @@ public class Listener implements ServletContextListener {
     public void contextInitialized(ServletContextEvent sce) {
         System.out.println("Hello Listener is works!");
         SqlHelper.initDB();
-
     }
 
     @Override
-    public void contextDestroyed(ServletContextEvent sce) {
-
-
-    }
+    public void contextDestroyed(ServletContextEvent sce) {}
 }
