+++ homework4_1/src/main/java/com/sergio/service/ProductService.java
@@ -1,0 +1,12 @@
+package com.sergio.service;
+
+import com.sergio.repository.ProductRepository;
+import java.util.Map;
+
+public class ProductService {
+
+    public static Map<String, Double> getAllProducts(){
+        Map<String, Double>products = ProductRepository.getProducts();
+        return products;
+    }
+}
